Support hipblasLT (#144)
* Revert "Revert hipblaslt (#141)"

This reverts commit 007e472417a5e630793974e9d57733bc6b979a78.

* setup: check if torch.cuda is available before calling get_device_properties

In hipBLASlt_supported(), torch.cuda.get_device_properties is called in
an env where visiable GPUs are not present. This will return an error
"RuntimeError: No HIP GPUs are available" and the build stops.
Check if torch.cuda is available and then access torch.cuda properties.


* Fix build error


* Fix build error for hipblas path


* Add support to pass env variable for hipblasLT support


* Update README.md

* Hold off adding env variable for now


* Update README.md

* Make IS_HIPBLASLT_SUPPORTED always True


---------

+++ csrc/multi_tensor_apply.cuh
@@ -19,7 +19,7 @@ constexpr int depth_to_max_blocks[5] = {2560, 2560, 2560, 2560, 2560};
 template<int n> struct TensorListMetadata
 {
   void* addresses[n][depth_to_max_tensors[n-1]];
-  int sizes[depth_to_max_tensors[n-1]];
+  int64_t sizes[depth_to_max_tensors[n-1]];
   unsigned char block_to_tensor[depth_to_max_blocks[n-1]];
   int block_to_chunk[depth_to_max_blocks[n-1]]; // I fear this needs to be a full int.
   int start_tensor_this_launch;
